backend/pyproject.toml
[tool.poetry]
name = "hospital-next-js-django"
version = "1.0.0"
description = "backend hospital-next-js-django"
authors = ["TheWitcher1991 <alikzoy@gmail.com>"]
license = "GPL-3.0"
readme = "README.MD"
repository = "https://github.com/TheWitcher1991/hospital-next-js-django"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1"
django-cors-headers = "^4.4.0"
django-debug-toolbar = "^4.4.6"
django-environ = "^0.11.2"
django-filter = "^24.3"
django-redis = "^5.4.0"
django-redisboard = "^8.4.0"
djangorestframework = "^3.15.2"
celery = "^5.4.0"
flower = "^2.0.1"
pyjwt = "^2.9.0"
pillow = "^10.4.0"
gunicorn = "^23.0.0"
drf-spectacular = "^0.27.2"
psycopg2 = "^2.9.9"
yookassa = "^3.3.0"
num2words = "^0.5.13"
django-prometheus = "^2.3.1"


[tool.poetry.group.dev.dependencies]
black = "^24.8.0"
flake8 = "^7.1.1"
isort = "^5.13.2"

[tool.black]
line-length = 120
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 120
multi_line_output = 3
include_trailing_comma = true
use_parentheses = true
ensure_newline_before_comments = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
